requirements.txt
flask
google-generativeai
langdetect
python-dotenv
